1. fix number of predictor training dataset; 2. improve trainning script 3. use dataset from two iteration to train seq2seq model

+++ baselines/baselines/ppo2/train_cycle.sh
@@ -7,11 +7,18 @@ function finish() {
         exit
 }
 
-rl_model="./log"
-pred_model="./pred"
+# preparation
+if [! -f "./models"]; then
+    mkdir "./models"
+else
+    rm -rf "./models/*"
+fi
 
-rm -rf "./models/*"
+rl_model="./log/"
+pred_model="./pred/"
 
+
+# start training
 counter=0
 save_counter=0
 while [ ${counter} -le ${1} ]
@@ -23,7 +30,7 @@ if [ ${counter} -eq 0 ]
 then
     python run.py -t=True --pred_weight=0.0
 else
-    python run.py -t=True -l=True -p='00200' --pred_weight=${2}
+    python run.py -t=True -l=True -p='00050' --pred_weight=${2}
 fi
 
 
@@ -31,18 +38,12 @@ fi
 sleep 1
 
 # sample dataset
-python run.py -l=True -p='00200'
+python run.py -l=True -p='00050'
 sleep 1
 
 # train seq2seq
 python predictor.py -l=True --iter=${counter}
 
-# if [ ${counter} -eq 0]
-# then
-# 	python predictor.py --iter=${counter}
-# else
-# 	python predictor.py -l=True --iter=${counter}
-# fi
 
 # copy saved file and rename
 cp -R ${rl_model} "./models/log_${counter}"
@@ -51,7 +52,6 @@ cp -R ${pred_model} "./models/pred_${counter}"
 rm -rf "${pred_model}/test1/checkpoint_" 
 
 ((counter++))
-
 done
 
 echo All done
